correction

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -11,20 +11,20 @@
 
 int main(void)
 {
-	int fib1 = 1;
-	int fib2 = 2;
-	int i;
+	unsigned long fib1 = 1;
+	unsigned long fib2 = 2;
+	unsigned long i;
 
-	printf("%d, %d", fib1, fib2);
+	printf("%lu, %lu", fib1, fib2);
 
 	for (i = 3; i <= 50; i++)
 	{
-		int currentfib = fib1 + fib2;
+		unsigned long currentfib = fib1 + fib2;
 
 		if (i < 50)
-			printf(", %d", currentfib);
+			printf(", %lu", currentfib);
 		else
-			printf(", %d\n", currentfib);
+			printf(", %lu\n", currentfib);
 
 		fib1 = fib2;
 		fib2 = currentfib;
